Consolidate inductor FX node utilities into fx_utils
Move general-purpose FX node/graph manipulation functions from
pattern_matcher.py, post_grad.py, and split_cat.py into fx_utils.py
so they can be reused without pulling in the full pattern matcher
module. The moved functions are re-exported from their original
locations for backward compatibility.

Also update joint_graph.py to use is_mutation_op from fx_utils
directly instead of a local is_mutated helper.

Authored with Claude.

[ghstack-poisoned]
